check namespaces
