+++ tools/rsdl/dotnet/rapid.rdm/RdmService.cs
@@ -1,0 +1,43 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace rapid.rdm
+{
+    /// <summary>
+    /// interfaces can't be sealed but remainder assumes only two implementations: RdmServiceCollection, RdmServiceSingelton
+    /// </summary>
+    public interface IRdmServiceElement
+    {
+        string Name { get; }
+    }
+
+    public class RdmService : IRdmSchemaElement, IEquatable<RdmService>
+    {
+        public RdmService()
+        {
+        }
+
+        public RdmService(IEnumerable<IRdmServiceElement> items)
+        {
+            Items = items;
+        }
+
+        public IEnumerable<IRdmServiceElement> Items { get; }
+
+        public bool Equals(RdmService other)
+        {
+            return Enumerable.SequenceEqual(this.Items, other.Items);
+        }
+
+        public override bool Equals(object other)
+        {
+            return other is RdmService model && this.Equals(model);
+        }
+
+        public override int GetHashCode()
+        {
+            return HashCode.Combine(Items);
+        }
+    }
+}
