+++ tools/rsdl/dotnet/rapid.rsdl.tests/ParserTests.cs
@@ -9,13 +9,13 @@ namespace rapid.rsdl.tests
         [Fact]
         public void TypePropertiesGetParsed()
         {
-            var content = "type company { name: string incorporated: date}";
+            var content = "type Company { name: String incorporated: Date}";
             var actual = RdmParser.Parse(content);
 
             var expected = new RdmDataModel(null, new[] {
-                new RdmStructuredType("company", new [] {
-                        new RdmProperty ("name", new RdmTypeReference("string"), null, new Position(1,16)),
-                        new RdmProperty ("incorporated", new RdmTypeReference("date"), null, new Position(1,29))
+                new RdmStructuredType("Company", new [] {
+                        new RdmProperty ("name", new RdmTypeReference("String"), null, new Position(1,16)),
+                        new RdmProperty ("incorporated", new RdmTypeReference("Date"), null, new Position(1,29))
                 })
             });
             Assert.Equal(expected, actual);
@@ -24,15 +24,83 @@ namespace rapid.rsdl.tests
         [Fact]
         public void NameSpaceDeclarationGetParsed()
         {
-            var content = "namespace foo.bar type company { }";
+            var content = "namespace foo.bar type Company { }";
             var actual = RdmParser.Parse(content);
 
             var expected = new RdmDataModel(new RdmNamespaceDeclaration("foo.bar"),
                 new[] {
-                    new RdmStructuredType("company", new RdmProperty[] {
+                    new RdmStructuredType("Company", new RdmProperty[] {
                 })
             });
 
+            Assert.Equal(expected, actual);
+        }
+
+        [Fact]
+        public void NamespaceReferencesGetParsed()
+        {
+            var content = @"
+namespace foo.bar
+include other.namespace as other from ""other.rsdl""
+type Company { }";
+            var actual = RdmParser.Parse(content);
+
+            var expected = new RdmDataModel(
+                new RdmNamespaceDeclaration("foo.bar"),
+                new[] {
+                    new RdmStructuredType("Company", new RdmProperty[] {})
+                },
+                new[] {
+                    new RdmNamespaceReference("other.namespace", "other", "other.rsdl")
+                }
+            );
+
+            Assert.Equal(expected.References, actual.References);
+            Assert.Equal(expected, actual);
+        }
+
+        [Fact]
+        public void NamespaceReferencesWithoutAliasGetParsed()
+        {
+            var content = @"
+namespace foo.bar
+include other.namespace from ""other.rsdl""
+type Company { }";
+            var actual = RdmParser.Parse(content);
+
+            var expected = new RdmDataModel(
+                new RdmNamespaceDeclaration("foo.bar"),
+                new[] {
+                    new RdmStructuredType("Company", new RdmProperty[] {})
+                },
+                new[] {
+                    new RdmNamespaceReference("other.namespace", null, "other.rsdl")
+                }
+            );
+
+            Assert.Equal(expected.References, actual.References);
+            Assert.Equal(expected, actual);
+        }
+
+        [Fact]
+        public void QualifiedTypeNamesGetParsed()
+        {
+            var content = @"
+include other.namespace as other from ""other.rsdl""
+type Company { something: other.Something }";
+            var actual = RdmParser.Parse(content);
+
+            var expected = new RdmDataModel(
+                null,
+                new[] {
+                    new RdmStructuredType("Company", new [] { new RdmProperty("something", new RdmTypeReference("other.Something"))})
+                },
+                new[] {
+                    new RdmNamespaceReference("other.namespace", "other", "other.rsdl")
+                }
+            );
+
+            Assert.Equal(expected.References, actual.References);
             Assert.Equal(expected, actual);
         }
     }
